// frontend/src/generated/openapi/model/creacionIntercambio.ts
/**
 * amoxtli
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Informacińo requerida para crear un nuevo intercambio
 */
export interface CreacionIntercambio { 
    /**
     * ID del usuario que recibe la oferta de intercambio
     */
    idAceptante: number;
    /**
     * ID del libro que pertenece al usuario que recibe la oferta de intercambio
     */
    idLibroAceptante: number;
}